refactor: RockPaperScissorsGame struct -> class로 변경
