+++ RockPaperScissors/RockPaperScissors/User.swift
@@ -1,0 +1,12 @@
+//
+//  User.swift
+//  RockPaperScissors
+//
+//  Created by 조민호 on 2022/02/15.
+//
+
+import Foundation
+
+struct User {
+
+}
